search_app/requirements.txt
blinker==1.9.0
certifi==2025.4.26
click==8.1.8
elastic-transport==8.17.0
elasticsearch==8.17.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
neo4j==5.28.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
python-dotenv

